requirements.txt
Django==1.10
django-filter==1.1.0
MySQL-python==1.2.5
pytz==2018.5
